Fix build for non-Unix systems
When we wrote the initial version of the crate we did not have the means
available to build it on non-Unix systems and so we had to eyeball the
code. That backfired royally.
With this change we fix up the build on non-Unix systems, on which we
can now build.