Use clippy lint for flagging usage of print & println
So far our stance was to have the user pass in an object implementing
Write that is to be used for all things output. With upcoming changes we
are about to revisit this decision and intend to use stdout directly
instead. Once we do that, our print and println overrides may flag
usage of the macros of the same name.
This change replaces our overrides with the usage of clippy lints that
flag similar usage. Usage of lints is much more flexible, because they
can be disabled much more easily.