Support handling additional sources in print_rerun_if_changed()
As it turns out, the get_revsision() function's signature is not quite
amenable to allow clients to use it in a completely reliable way.
Specifically, the function emits information to rebuild the build script
if some internal git files were changed. However, in order to correctly
capture local modifications, we actually also need to rebuild the script
when sources of the program change. In the general case, this crate
cannot know what these sources are, meaning that they will ultimately
need to be user provided.
As a first step towards allowing users to provide sources as input, this
change adjusts the print_rerun_if_changed() function to accept
additional sources to emit 'cargo:rerun-if-changed' lines for.