Print cargo:rerun-if-changed directives with absolute path
Currently we print the cargo:rerun-if-changed directives that Cargo
requires with a path relative to what is assumed to be the project root
that Cargo uses. There are many directory configurations that Git
supports and with Cargo workspaces, probably similarly many
possibilities on the Cargo side of things. It is thus conceivable that
the .git directory is not actually located at a location that Cargo
interprets this directives to be relative to.
To circumvent all this uncertainly, this change makes us work with
absolute paths instead.