Introduce with_valid_git() function
With more and more functions added to the public API surface, it becomes
increasingly bad design to funnel all code paths to a single function,
revision_bare_impl(). This function will accumulate complexity as we
necessarily have to request it to perform more jobs.
This change re-envisions what revision_bare_impl() should take care of
by factoring out the Git validity checks into a new function,
with_valid_git(). The assumption is that this function is called as
pretty much the first thing in every public function. Right now there is
no enforcement of that happening, as overall complexity is still tiny,
but we can always introduce some kind of token mechanism that enforces
this constraint more strictly.