Introduce git_revision_auto() function
Mulling over the required correct set of rebuild directives to emit some
more, we have come to the conclusion that not we can actually determine
the set of source files that we care about changing programmatically.
Namely, given that any modification of any file tracked by Git could
cause a change of the produced revision string (by adding the '+' marker
indicating local changes), the build script effectively depends on all
tracked git files. That is to say, the sources argument to the
git_revision() function is unnecessary at a conceptual level, and a
potential source of headaches and problems for clients.

This change addresses this issue by introducing git_revision_auto().
This function is very much similar to the previously added
git_revision(), but it determines the set of source files to emit
cargo:rerun-if-changed directives for programmatically, by querying Git.